update rust code and cpp files location
