+++ SConstruct
@@ -26,6 +26,10 @@ sources = Glob("src/*.cpp")
 if env["platform"] == "windows":
     env.Append(CPPDEFINES="WINDOWS_ENABLED")
 
+if env["precision"] == "double":
+	env.Append(CPPDEFINES=["REAL_T_IS_DOUBLE"])
+else:
+	env.Append(CPPDEFINES=["REAL_T_IS_FLOAT"])
 if env["platform"] == "macos":
 	library = env.SharedLibrary(
 		"bin/addons/godot-rapier2d/bin/libphysics_server_rapier2d.{}.{}.framework/libphysics_server_rapier2d.{}.{}".format(
