+++ src/rapier2d-wrapper/cbindgen.toml
@@ -14,10 +14,10 @@ language = "C++"
 
 # header = "/* Text to put at the beginning of the generated file. Probably a license. */"
 # trailer = "/* Text to put at the end of the generated file */"
-# include_guard = "my_bindings_h"
-pragma_once = true
+include_guard = "RAPIER_WRAPPER_H"
+#pragma_once = true
 # autogen_warning = "/* Warning, this file is autogenerated by cbindgen. Don't modify this manually. */"
-include_version = false
+include_version = true
 namespace = "rapier2d"
 namespaces = []
 using_namespaces = []
@@ -53,8 +53,10 @@ usize_is_size_t = true
 [defines]
 # "target_os = freebsd" = "DEFINE_FREEBSD"
 # "feature = serde" = "DEFINE_SERDE"
-
-
+"feature = f32" = "REAL_T_IS_FLOAT"
+#"feature = f64" = "REAL_T_IS_FLOAT"
+# Not needed
+"feature = spin_no_std" = "DEFINE_SPIN_NO_STD"
 
 [export]
 include = []
@@ -137,7 +139,7 @@ bitflags = false
 ############## Options for How Your Rust library Should Be Parsed ##############
 
 [parse]
-parse_deps = false
+parse_deps = true
 # include = []
 exclude = []
 clean = false
