+++ src/shapes/rapier_shape_2d.cpp
@@ -90,19 +90,19 @@ Variant RapierWorldBoundaryShape2D::get_data() const {
 /*********************************************************/
 /*********************************************************/
 
-// void RapierSeparationRayShape2D::set_data(const Variant &p_data) {
-// 	Dictionary d = p_data;
-// 	length = d["length"];
-// 	slide_on_slope = d["slide_on_slope"];
-// 	configure(Rect2(0, 0, 0.001, length));
-// }
+void RapierSeparationRayShape2D::set_data(const Variant &p_data) {
+	Dictionary d = p_data;
+	length = d["length"];
+	slide_on_slope = d["slide_on_slope"];
+	configure(Rect2(0, 0, 0.001, length));
+}
 
-// Variant RapierSeparationRayShape2D::get_data() const {
-// 	Dictionary d;
-// 	d["length"] = length;
-// 	d["slide_on_slope"] = slide_on_slope;
-// 	return d;
-// }
+Variant RapierSeparationRayShape2D::get_data() const {
+	Dictionary d;
+	d["length"] = length;
+	d["slide_on_slope"] = slide_on_slope;
+	return d;
+}
 
 /*********************************************************/
 /*********************************************************/
